+++ src/index.js
@@ -18,6 +18,8 @@ import StepperComponent from '@/components/stepper/StepperComponent.vue';
 import ToolTip from '@/components/tooltip/ToolTip.vue';
 import CalendarComponent from '@/components/calendar/CalendarComponent.vue';
 import NoContent from '@/components/nocontent/NoContent.vue';
+import NotificationComponent from '@/components/notification/NotificationComponent.vue';
+import NotificationPlugin from './plugins/NotificationPlugin';
 
 const GtsPrint = {
   install(app) {
@@ -40,6 +42,8 @@ const GtsPrint = {
     app.component('ToolTip', ToolTip);
     app.component('CalendarComponent', CalendarComponent);
     app.component("NoContent", NoContent);
+    app.component("NotificationComponent", NotificationComponent);
+    app.use(NotificationPlugin);
   }
 };
 
@@ -63,7 +67,9 @@ export {
   StepperComponent,
   ToolTip,
   CalendarComponent,
-  NoContent
+  NoContent,
+  NotificationComponent,
+  NotificationPlugin 
 };